enforce sizes in archive formats

+++ TheForceEngine/TFE_Archive/gobArchive.h
@@ -48,19 +48,19 @@ private:
 	typedef struct
 	{
 		char GOB_MAGIC[4];
-		long MASTERX;	// offset to GOB_Index_t
+		u32 MASTERX;	// offset to GOB_Index_t
 	} GOB_Header_t;
 
 	typedef struct
 	{
-		long IX;		// offset to the start of the file.
-		long LEN;		// length of the file.
+		u32 IX;		// offset to the start of the file.
+		u32 LEN;		// length of the file.
 		char NAME[13];	// file name.
 	} GOB_Entry_t;
 
 	typedef struct
 	{
-		long MASTERN;	// num files
+		u32 MASTERN;	// num files
 		GOB_Entry_t *entries;
 	} GOB_Index_t;
